Fix goroutine semantics bug in latch for post-flip signalling.
I wish I could think of a reasonable way to test or lint things like this.  It's one of those "what's the acceptable tree of folks to block?" things that we need so badly to provide consistent and well-documented answers to throughout this library.

I sort of aesthetically question/dislike using a single-use teeny goroutine like this, but clearly keeping a long lived one for arbitrarily distant future closes is far more absurd, so, here we go.

Signed-off-by: Eric Myhre <hash@exultant.us>