Okay, I think I know how I want to define supervisor lifecycle: More Thunks.
Using a callback to define the scope and lifecycle of the supervisor control itself makes a lot of sense: it aligns the stack frames with the reality of your control needs so you never have to remember to call any ugly shutdown methods (not even defer them), and all your worst-case error handling (something we're deeply emotional about for the purposes of this library) can be fenced in last-ditch defenses by the sup library itself.

The distinction between a root supervisor and just another supervisor down the tree instantly looked silly in this light so it's gone.  I'm wondering if this is going to make me feel like the distinction between Witness|Chaperon and Supervisor are silly shortly down the road as well, but holding off on that for now... I'd kind of like to think it's reasonably obvious when you want a supervisor vs a single leaf process that will now be linear.  But that's... yeah, that sounding less realistic per second that passes.

Signed-off-by: Eric Myhre <hash@exultant.us>