Implement second "worker" in the example, which is itself another supervisor.
This has some fixmes accumulating: This code is a great example of why passing down a quit chan Isn't Enough; if you forget that snippet at the end to forward quits, the example will hang!

... Which, of course, I immediately did to myself during development.

So, we need to architect until that Goes Away.  And I think this example provides pretty good evidence that having the concepts of chaparones and supervisors be disjoint is turning out unhelpful there: supervisors should be able to pass messages down between their own known-to-be-well-behaved internal "secretary" actors.

Signed-off-by: Eric Myhre <hash@exultant.us>