I need some Things to help test situations where events *correctly* have multiple valid orderings because the operations are intentionally nonlinearizable.
This is aiming for "simplest thing that works".  A "better" solution probably approaches the same complexity as a regexp engine or a lexer state machine, which... while being more correct for a more complex category of inputs, would be proportionally less informative when it failed.  So, this: a series of flat sequence checks, going for easy to write and clear to debug.

Note that this also doesn't help in the slighest at *confirming* that multiple routes may be taken.  That would be... possible, but not the greatest use of CPU time.  Instead, it's better to plan other tests which force concurrent operations and would either immediately fail (or at least clearly deadlock) so that we can complete tests without burning silicon.

Signed-off-by: Eric Myhre <hash@exultant.us>