Oh stats
I'm going to have to resist the temptation to build a better mousetrap for benchmarking as a result of some of this.  The answer in most serious situations is "don't freakin try to microbench so hard", but like... the JVM, with all its vaunted horrors when it comes to benchmarking, can also be hammered into statistical submission down to the nanos.  I want that for go.  (And specifically, for some of the same reasons: the paper about the JVM I'm visualizing (but can't find/recall a link for) was going full ham on synchronization overheads under various contention scenarios.  In other words -- literally this.)

Signed-off-by: Eric Myhre <hash@exultant.us>