Tests of more parts of the breakdown.
Also, start keeping memory stats.

Also, rename the funcs to refer to gathering, which is the mode of use we're exploring right now, but may not be the only one in the future.

Results from another run were interesting; they nearly sum up as desired, but there's still a little slippage:

	(trigr+subsc) = sum(-obsrv) = unexplained
	  191 +    0  = 191  -192   = -1
	  257 +  733  = 990  -913   = 77
	  320 + 1503  = 1823 -1753  = 70
	  461 + 3481  = 3942 -3484  = 458
	  723 + 6856  = 7579 -7256  = 323

	so they're all slightly slower when split, and it gets worse with more chans?  seems legit
	5~13%, though.  (...13?  I don't know why this appears to be a step function.)

The fractionated tests also incur the memory stats conversations more, by virtue of more start/stop calls, but those should be O(1) and not moving with our num-chans parameter.

Signed-off-by: Eric Myhre <hash@exultant.us>