Bench for unbuffered channels with latches!  These are more comparable to the behavior of the "fuse" implementation, and lo: channel sends and channel closes are not all created equal!  Behavior on the far side *matters*.
So, though this was quite the journey... this is why it was worth it to refrain from building a whole system to test at the start.  Sure, real systems are going to face contention issues when understanding performance.  But boy, is it useful to know that there's drastically different behavioral modes in the runtime for some of these very fundamental operations.

Signed-off-by: Eric Myhre <hash@exultant.us>