Discard chaperon and task: all is supervisor.
Fairly massive refactor to make all nodes and leaves into a single supervisor type.  The results seem much more reasonable to work with.  (Aside the fact there's more secretaries than perhaps necessary.)

And more things are interfaces, because I see no reason to expose users to the fact that you need a pointer to your supervisor implementation.

I'm having a nasty argument with the race detector, though.  It's flagging a bunch of things that are interacting with... os.Stdout.  Current theory: somewhere in the refactor I made it more possible for grandchildren to outlive the root supervisor, and that's causing flack, but no tests go deterministically wrong from this (yet/unfortunately).

Signed-off-by: Eric Myhre <hash@exultant.us>