Update saltmines example:
- remove extraneous supervisor (now that there is no distinction between those and chaperons)
- yank the ore washer workers to their own func, instead of leaving them anon
- remove manual cancellation passdown from the owf director, because that's suppose to work now

That last one also appears to fix a live lock which... it shouldn't, and I can't figure out for the life of me.  The quit certainly is sent by the overseer func.  And the disbatch down doesn't block in any way.  But we need more specific tests to chase whatever this is, I guess.

Signed-off-by: Eric Myhre <hash@exultant.us>