Port saltmines example to new api.
Unintending a massive block of code that ended up in an anonymous func feels good.  Not having the witness objects to hurl around feels ok but only because I wasn't using them for anything useful (I think that's going to remain defensible though).

The test is overall in need of More -- there's not enough of it done to really verify much about the series of steps we're hitting; only to roughly confirm whether we successfully avoid dead-(or live-)lock -- but at least it's as functional as it was before.

Signed-off-by: Eric Myhre <hash@exultant.us>