Work through another example where an agent is itself a long lived actor that commands a manager to spawn new parallel running agents.
We're missing the mark on some of the error handling responsiveness; that can be improved shortly.  This example was to research what API we need: and it's pretty clear you should be able to get a selectable channel from a manager to pick up any any errors when you're an actor.  Tis still a little tricky, there, though: can we keep refraining from needing a secretary thread to service such a channel? and also avoid making it dangerously blocky and deadlock prone for returning child funcs upsubbing their errors?

Also learned: It's a littlle dangerous to spawn an infinite number of goroutines like this.  If you typo that channel to keep receiving zeros after it's closed, whoooweeee you're gonna be sad.  You'll probably OoM; or, perhaps even more annoyingly, at some point you get killed and the attempt to serialize 300k goroutine stacks has your terminal flowing at warp speed for a couple minutes straight.  We should probably have a LeakyBucket sort of behavior helper for this, though that too is not a guaranteed panacea: an infinite queue there is perhaps a constant factor less alarming in memory use, and less irritating to read stacks from if it topples, but still fundamentally a route to oom if you screw up your ingress too badly and/or fail at backpressure.

Signed-off-by: Eric Myhre <hash@exultant.us>