Implement a 'canal' type for bilaterally nonblocking handoffs.
At the cost of unbounded memory costs, sure.  There really are times when I'm OK with that, though.  For example, I'm thinking this might be useful for buffering the error/exit reports from child procs: definitely don't wanna drop em; badly need the submitter to not ever block; and needs selectability on the reader side.  The next closest fit would be a channel with a "this is high enough" capacity.  This is far less kludgy.

"Canal" because it's a like a channel, but doesn't really care if the waters get high and aren't flowing particularly fast.  Name subject to change though.

Signed-off-by: Eric Myhre <hash@exultant.us>