Rn 'canal'->'sluice'.
Harder to spell.  But an accurate metaphor: it's a bit like a channel, but sometimes you just don't really care for it to flow right now.  (Will be even more apt when I get around to adding a version that accepts a pair of functions for internal-buffer-append and decide-whats-next instead of just this basic queueing.)

Signed-off-by: Eric Myhre <hash@exultant.us>