Fix a sluice test with probablistic failure.
The synchronization wasn't wrong: the test was.  If you sign up another channel first, *then* send, it's random -- up to map order -- which one will get it.

The test now sends that message, then signs up the next channel, making the behavior predictable.

Though it's always amusing to me to see things like map iteration on a 2-item map be "AB" 80% of the time and "BA" 20% of the time.  An interesting form of random, that.  Nobody promised uniform distribution, just... wat.

Signed-off-by: Eric Myhre <hash@exultant.us>