Rn `NewWrit`->`NewTask`, and document it.
There's some dissonance with naming here between Writ and Task; I guess it makes sense as long as you imagine that "task" refers to the yin-yang of the writ and its paired supervisor.

At any rate, this function should clearly have the same name as its equivalent method on Manager, and now it does.

Signed-off-by: Eric Myhre <hash@exultant.us>