Attempt to log slow quitters from manager.Work.
This is now... quite the state machine itself, implemented in a breakneck tangle of gotos, and doing an awfully large amount of thinking for something that was meant to just be a helpful default "park" function.  I don't yet have usecases that suggest any obviously (ha) reasonable way to break things down, though.

Signed-off-by: Eric Myhre <hash@exultant.us>