Link meep.
Considered avoiding this as a dependency by doing something similar to the approach for logs: making an exported function reference you can set to control error handling.  Decided against that, and to take a strong stance using meep's error system, because fragmentation on the part of different applications would make it nearly impossible for libraries to know how to handle errors, and because something that captures stacks from the originating goroutine is pretty much a hard requirement for usability (an opinionated statement, but one I feel pretty okay making).

Signed-off-by: Eric Myhre <hash@exultant.us>