Add "important" bool to log func, hinting whether it's highly advisible to print a message or not.
Logging go-sup events can cause serious log fatigue in large applications (and often even very *small* applications if they use the network or other heavy triggers of concurrency).  So it's been tempting to put them behind "--verbose" flags in end-user applications.  But this is not ideal if there *is* a programmer error of some kind which, for example, caused one actor to be unable to respond to quits properly, resulting in a systemic hang -- we have helpful warnings for these situations, and it's *really* important to be able to see and gather those.

So, here's our compromise: we'll indicate whether a message is highly important.  go-sup still won't import any particular logging framework as an API dependency, but if you use a framework, this bool probably translates to whether you should log the message as a "debug" level or a "warn" level.

Signed-off-by: Eric Myhre <hash@exultant.us>