Drop TraitTraceable from ErrTaskPanic; make it record its task name instead.
The stack trace of the area where the ErrTaskPanic is created seems to fairly irrelevant: the original error should contain a stacktrace if it opted into one, and every additional manager layer above this that bubbles up a ErrTaskPanic tends to contribute a stack parked in `go-sup.manager.Work()`, which is fairly useless in its redundancy.

There's an argument to be made for the middle area of the stacks containing some information on *how* the parent agent func trees got down to the `Work()` call, and maybe in the future we can come up with some compromise to get the best of that extracted, but so far the practical utility of retaining the whole stack has been... pretty much net negative, considering how much sheer noise it adds vs signal.

Adding the task name should be pretty uncontrovertial.  That's the one thing everyone looks for first when reading a tree of these.

Signed-off-by: Eric Myhre <hash@exultant.us>